requirements.txt
# core
requests
rich
python-box
pydantic
ruamel.yaml
pyyaml
# docs
sphinx
sphinx-rtd-theme
sphinx-autodoc-typehints
